Update GApps ZIP_MOVE logic

+++ GApps/gapps.sh
@@ -103,12 +103,16 @@ if [[ `ls ${SOURCE_DIR}/out/${ZIP_BEG}*.zip 2>/dev/null | wc -l` != "0" ]]; then
    BUILD_RESULT_STRING="BUILD SUCCESSFUL"
    SUCCESS=true
 
-
-   # Remove current GApps and move the new ones in their place
-   if [[ "${TYPE}" == "banks" && ${PERSONAL} = false ]]; then
-      rm -rf ${HOME}/shared/.me/${ZIP_BEG}*.zip
+   # If ZIP_MOVE doesn't exist, make it; otherwise, clean it
+   if [[ ! -d ${ZIP_MOVE} ]]; then
+      mkdir -p ${ZIP_MOVE}
+   else
+      # Remove current GApps and move the new ones in their place
+      if [[ "${TYPE}" == "banks" && ${PERSONAL} = false ]]; then
+         rm -rf ${HOME}/shared/.me/${ZIP_BEG}*.zip
+      fi
+      rm -rf ${ZIP_MOVE}/${ZIP_BEG}*.zip
    fi
-   rm -rf ${ZIP_MOVE}/${ZIP_BEG}*.zip
 
    if [[ "${TYPE}" == "banks" && ${PERSONAL} = false ]]; then
       cp -v ${SOURCE_DIR}/out/${ZIP_BEG}*.zip ${HOME}/shared/.me
